+++ listeners/views/index.js
@@ -1,4 +1,3 @@
-import { addCommentSubmitCallback } from "./add-comment-submit.js";
 import { retroSubmitCallback } from "./retro-submit.js";
 
 /**
@@ -7,5 +6,4 @@ import { retroSubmitCallback } from "./retro-submit.js";
  */
 export const register = (app) => {
   app.view("retro_submit", retroSubmitCallback);
-  app.view("add_comment_submit", addCommentSubmitCallback);
 };